Fix device_id

+++ iot/startup.sh
@@ -2,11 +2,21 @@
 
 echo "***** STARTUP *****"
 
-kill $(ps -ef | grep rawprinter | head -1 | awk '{print $2}')
-kill $(ps -ef | grep serialscale | head -1 | awk '{print $2}')
-kill $(ps -ef | grep minirps | head -1 | awk '{print $2}')
+if [[ $(ps -ef | grep rawprinter | wc -l) -gt 1 ]]; then
+  kill $(ps -ef | grep rawprinter | head -1 | awk '{print $2}')
+fi
 
-rawprinter --vendor-id 0x0a5f --device-id 0x000a &
+if [[ $(ps -ef | grep serialscale | wc -l) -gt 1 ]]; then
+  kill $(ps -ef | grep serialscale | head -1 | awk '{print $2}')
+fi
+
+if [[ $(ps -ef | grep minirps | wc -l) -gt 1 ]]; then
+  kill $(ps -ef | grep minirps | head -1 | awk '{print $2}')
+fi
+
+ID=$(lsusb | grep -w Zebra | awk '{print $6}' | awk -F ':' '{print$2}')
+
+rawprinter --vendor-id 0x0a5f --device-id 0x$ID &
 serialscale /dev/ttyUSB0 --lang pt --unit Kg --min-weight 0.01 &
 minirps -f /etc/iot/config.toml &
